+++ arviz/plots/backends/matplotlib/densityplot.py
@@ -3,7 +3,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from . import backend_show
-from ....stats import hpd
+from ....stats import hdi
 from ...plot_utils import (
     make_label,
     _create_axes_grid,
@@ -26,9 +26,9 @@ def plot_density(
     xt_labelsize,
     linewidth,
     markersize,
-    credible_interval,
+    hdi_prob,
     point_estimate,
-    hpd_markers,
+    hdi_markers,
     outline,
     shade,
     n_data,
@@ -64,9 +64,9 @@ def plot_density(
                 xt_labelsize,
                 linewidth,
                 markersize,
-                credible_interval,
+                hdi_prob,
                 point_estimate,
-                hpd_markers,
+                hdi_markers,
                 outline,
                 shade,
                 axis_map[label],
@@ -92,9 +92,9 @@ def _d_helper(
     xt_labelsize,
     linewidth,
     markersize,
-    credible_interval,
+    hdi_prob,
     point_estimate,
-    hpd_markers,
+    hdi_markers,
     outline,
     shade,
     ax,
@@ -121,8 +121,8 @@ def _d_helper(
         Thickness of lines
     markersize : float
         Size of markers
-    credible_interval : float
-        Credible intervals. Defaults to 0.94
+    hdi_prob : float
+        hdi intervals. Defaults to 0.94
     point_estimate : Optional[str]
         Plot point estimate per variable. Values should be 'mean', 'median', 'mode' or None.
         Defaults to 'auto' i.e. it falls back to default set in rcParams.
@@ -132,14 +132,14 @@ def _d_helper(
     ax : matplotlib axes
     """
     if vec.dtype.kind == "f":
-        if credible_interval != 1:
-            hpd_ = hpd(vec, credible_interval, multimodal=False)
-            new_vec = vec[(vec >= hpd_[0]) & (vec <= hpd_[1])]
+        if hdi_prob != 1:
+            hdi_ = hdi(vec, hdi_prob, multimodal=False)
+            new_vec = vec[(vec >= hdi_[0]) & (vec <= hdi_[1])]
         else:
             new_vec = vec
 
         density, xmin, xmax = _fast_kde(new_vec, bw=bw)
-        density *= credible_interval
+        density *= hdi_prob
         x = np.linspace(xmin, xmax, len(density))
         ymin = density[0]
         ymax = density[-1]
@@ -153,16 +153,16 @@ def _d_helper(
             ax.fill_between(x, density, color=color, alpha=shade)
 
     else:
-        xmin, xmax = hpd(vec, credible_interval, multimodal=False)
+        xmin, xmax = hdi(vec, hdi_prob, multimodal=False)
         bins = get_bins(vec)
         if outline:
             ax.hist(vec, bins=bins, color=color, histtype="step", align="left")
         if shade:
             ax.hist(vec, bins=bins, color=color, alpha=shade)
 
-    if hpd_markers:
-        ax.plot(xmin, 0, hpd_markers, color=color, markeredgecolor="k", markersize=markersize)
-        ax.plot(xmax, 0, hpd_markers, color=color, markeredgecolor="k", markersize=markersize)
+    if hdi_markers:
+        ax.plot(xmin, 0, hdi_markers, color=color, markeredgecolor="k", markersize=markersize)
+        ax.plot(xmax, 0, hdi_markers, color=color, markeredgecolor="k", markersize=markersize)
 
     if point_estimate is not None:
         est = calculate_point_estimate(point_estimate, vec, bw)