+++ arviz/plots/backends/matplotlib/posteriorplot.py
@@ -5,7 +5,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from . import backend_show
-from ....stats import hpd
+from ....stats import hdi
 from ...kdeplot import plot_kde
 from ...plot_utils import (
     make_label,
@@ -29,7 +29,7 @@ def plot_posterior(
     kind,
     point_estimate,
     round_to,
-    credible_interval,
+    hdi_prob,
     multimodal,
     ref_val,
     rope,
@@ -63,7 +63,7 @@ def plot_posterior(
             kind=kind,
             point_estimate=point_estimate,
             round_to=round_to,
-            credible_interval=credible_interval,
+            hdi_prob=hdi_prob,
             multimodal=multimodal,
             ref_val=ref_val,
             rope=rope,
@@ -91,7 +91,7 @@ def _plot_posterior_op(
     bins,
     kind,
     point_estimate,
-    credible_interval,
+    hdi_prob,
     multimodal,
     ref_val,
     rope,
@@ -198,11 +198,9 @@ def _plot_posterior_op(
     def display_hpd():
         # np.ndarray with 2 entries, min and max
         # pylint: disable=line-too-long
-        hpd_intervals = hpd(
-            values, credible_interval=credible_interval, multimodal=multimodal
-        )  # type: np.ndarray
+        hdi_probs = hdi(values, hdi_prob=hdi_prob, multimodal=multimodal)  # type: np.ndarray
 
-        for hpdi in np.atleast_2d(hpd_intervals):
+        for hpdi in np.atleast_2d(hdi_probs):
             ax.plot(
                 hpdi,
                 (plot_height * 0.02, plot_height * 0.02),
@@ -227,7 +225,7 @@ def _plot_posterior_op(
             ax.text(
                 (hpdi[0] + hpdi[1]) / 2,
                 plot_height * 0.3,
-                format_as_percent(credible_interval) + " HPD",
+                format_as_percent(hdi_prob) + " HPD",
                 size=ax_labelsize,
                 horizontalalignment="center",
             )
@@ -271,7 +269,7 @@ def _plot_posterior_op(
     plot_height = ax.get_ylim()[1]
 
     format_axes()
-    if credible_interval is not None:
+    if hdi_prob != "hide":
         display_hpd()
     display_point_estimate()
     display_ref_val()