+++ arviz/plots/hpdplot.py
@@ -3,15 +3,16 @@ import numpy as np
 from scipy.interpolate import griddata
 from scipy.signal import savgol_filter
 
-from ..stats import hpd
+from ..stats import hdi
 from .plot_utils import get_plotting_function, matplotlib_kwarg_dealiaser
 from ..rcparams import rcParams
+from ..utils import credible_interval_warning
 
 
 def plot_hpd(
     x,
     y,
-    credible_interval=None,
+    hdi_prob=None,
     color="C1",
     circular=False,
     smooth=True,
@@ -22,9 +23,10 @@ def plot_hpd(
     backend=None,
     backend_kwargs=None,
     show=None,
+    credible_interval=None,
 ):
     r"""
-    Plot hpd intervals for regression data.
+    Plot hdi intervals for regression data.
 
     Parameters
     ----------
@@ -32,8 +34,8 @@ def plot_hpd(
         Values to plot
     y : array-like
         values from which to compute the hpd. Assumed shape (chain, draw, \*shape).
-    credible_interval : float, optional
-        Credible interval to plot. Defaults to 0.94.
+    hdi_prob : float, optional
+        HDI interval to plot. Defaults to 0.94.
     color : str
         Color used for the limits of the HPD interval and fill. Should be a valid matplotlib color
     circular : bool, optional
@@ -59,11 +61,16 @@ def plot_hpd(
         check the plotting method of the backend.
     show : bool, optional
         Call backend show function.
+    credible_interval: float, optional
+        deprecated: Please see hdi_prob
 
     Returns
     -------
     axes : matplotlib axes or bokeh figures
     """
+    if credible_interval:
+        hdi_prob = credible_interval_warning(credible_interval, hdi_prob)
+
     plot_kwargs = matplotlib_kwarg_dealiaser(plot_kwargs, "plot")
     plot_kwargs.setdefault("color", color)
     plot_kwargs.setdefault("alpha", 0)
@@ -87,13 +94,13 @@ def plot_hpd(
         new_shape = tuple([-1] + list(x_shape))
         y = y.reshape(new_shape)
 
-    if credible_interval is None:
-        credible_interval = rcParams["stats.credible_interval"]
+    if hdi_prob is None:
+        hdi_prob = rcParams["stats.hdi_prob"]
     else:
-        if not 1 >= credible_interval > 0:
-            raise ValueError("The value of credible_interval should be in the interval (0, 1]")
+        if not 1 >= hdi_prob > 0:
+            raise ValueError("The value of hdi_prob should be in the interval (0, 1]")
 
-    hpd_ = hpd(y, credible_interval=credible_interval, circular=circular, multimodal=False)
+    hdi_ = hdi(y, hdi_prob=hdi_prob, circular=circular, multimodal=False)
 
     if smooth:
         if smooth_kwargs is None:
@@ -102,12 +109,12 @@ def plot_hpd(
         smooth_kwargs.setdefault("polyorder", 2)
         x_data = np.linspace(x.min(), x.max(), 200)
         x_data[0] = (x_data[0] + x_data[1]) / 2
-        hpd_interp = griddata(x, hpd_, x_data)
+        hpd_interp = griddata(x, hdi_, x_data)
         y_data = savgol_filter(hpd_interp, axis=0, **smooth_kwargs)
     else:
         idx = np.argsort(x)
         x_data = x[idx]
-        y_data = hpd_[idx]
+        y_data = hdi_[idx]
 
     hpdplot_kwargs = dict(
         ax=ax,